fix: flask application factory and blueprints
