+++ Makefile
@@ -67,7 +67,7 @@ run-dev:
 	${PYTHON} src/onetime/entrypoints/web/manage.py runserver
 
 run-flask:
-	${PYTHON} -m flask --app src.onetime.entrypoints.web_flask.app --debug run --host 0.0.0.0
+	${PYTHON} -m flask --app src.onetime.entrypoints.web_flask.application --debug run --host 0.0.0.0
 
 
 collectstatic:
