+++ src/onetime/entrypoints/web_flask/blueprints/views.py
@@ -1,7 +1,6 @@
-from dependency_injector.wiring import Provide
-from flask import render_template, request, url_for
+from dependency_injector.wiring import Provide, inject
+from flask import Blueprint, render_template, request, url_for
 
-from onetime.entrypoints.web_flask.app import app
 from onetime.entrypoints.web_flask.forms import SecretCreateForm
 from onetime.use_cases.exceptions import (
     SecretDataWasAlreadyConsumedException,
@@ -10,18 +9,30 @@ from onetime.use_cases.exceptions import (
 )
 from onetime.use_cases.manager import SecretAndUrlManager
 
+blueprint = Blueprint("secret", __name__, url_prefix="/")
 
-@app.route("/", methods=["GET", "POST"])
+
+@blueprint.after_request
+def add_security_headers(response):
+    response.headers["Strict-Transport-Security"] = "max-age=31536000"
+    # response.headers['Cache-Control'] = 'public, max-age=31536000'
+    response.headers["X-Frame-Options"] = "SAMEORIGIN"
+    response.headers["Content-Type"] = "nosniff"
+    return response
+
+
+@blueprint.route("/", methods=["GET", "POST"])
+@inject
 def index(
     secret_and_url_manager: SecretAndUrlManager = Provide["secret_and_url_manager"],
 ):
     form = SecretCreateForm()
     if request.method == "POST":
         form = SecretCreateForm(request.form)
-        if not form.validate_on_submit():
+        if form.validate_on_submit():
             data = form.secret.data
             uuid = secret_and_url_manager.generate_secret_and_url(data)
-            secret_url = url_for("secret", uuid=uuid, _external=True)
+            secret_url = url_for("secret.secret", uuid=uuid, _external=True)
             return render_template(
                 "onetimesecrets/index.html", form=form, secret_url=secret_url
             )
@@ -29,7 +40,8 @@ def index(
     return render_template("onetimesecrets/index.html", form=form)
 
 
-@app.route("/secret/<string:uuid>", methods=["GET", "POST"])
+@blueprint.route("/secret/<string:uuid>", methods=["GET", "POST"])
+@inject
 def secret(
     uuid: str,
     secret_and_url_manager: SecretAndUrlManager = Provide["secret_and_url_manager"],
